fix(db): GetInt32-on-bigint -> ::int cast on COUNT/SUM projections (audit Batch 1)
PostgreSQL COUNT(*) and SUM(integer) return bigint; reading them via Npgsql
reader.GetInt32 throws InvalidCastException at runtime. 8 projection sites
across 6 services (Backend Lead+Analytics, Outbound, Automation, Knowledge,
Appointments, Integrations) get an explicit ::int cast so the existing
GetInt32 reads are correct. Behavior-preserving: per-tenant COUNT/SUM values
are far below int.MaxValue. Decimal SUM columns (value_amount,
deflection_revenue -> GetDecimal) and CASE-division columns (-> GetDouble)
deliberately left uncast. Removes latent crashes on analytics/stats endpoints.

Codex gpt-5.5 iter1 PASS (12/12 CQ + 5/5 CoVe, 0 blocking). iter0 CQ9
(tenant-isolation in touched Appointments/Knowledge SQL) resolved with NO
source change via corrected-premise (projection-only cast leaves every tenant
predicate byte-identical; Appointments subquery correlated-safe to tenant-scoped
outer; wa_intents analysis_id-scoped is pre-existing, tenant-hardening tracked
as separate Batch 8) + arch/codex-context.md anti-false-positive entry.

No migration, no Invekto.Shared change, no deploy (Q: commit'te dur).
Refs: arch/reports/20260614-audit-INDEX.md Batch 1

+++ src/Invekto.Automation/Data/AutomationRepository.cs
@@ -1280,8 +1280,8 @@ public sealed class AutomationRepository
     {
         const string sql = @"
             SELECT
-                COUNT(*) AS total,
-                COUNT(*) FILTER (WHERE was_deflected = TRUE) AS deflected,
+                COUNT(*)::int AS total,
+                COUNT(*) FILTER (WHERE was_deflected = TRUE)::int AS deflected,
                 COALESCE(SUM(deflection_revenue) FILTER (WHERE was_deflected = TRUE), 0) AS revenue_saved,
                 reason_category,
                 action_taken
